change functionality of Auth to Modals
