+++ src/components/EmbeddedVideo.jsx
@@ -6,7 +6,7 @@ import ButtonAction from './ButtonAction';
 import '../styles/components/EmbeddedVideo.scss';
 import LikeIcon from '../assets/icons/LikeIcon';
 import ShareIcon from '../assets/icons/ShareIcon';
-import { showRegisterForm } from '../store/reducers/Auth.reducer';
+import { showRegisterForm } from '../store/reducers/Modals.reducer';
 
 const EmbeddedVideo = (props) => {
   const { data, subscribers, descriptions } = props;
